Move legacy files back to where they were, v2 to separate dropdown-menu-v2 folder, fix docs
