+++ docs/manifest.json
@@ -816,15 +816,9 @@
 		"parent": "components"
 	},
 	{
-		"title": "V1",
-		"slug": "v1",
-		"markdown_source": "../packages/components/src/dropdown-menu/v1/README.md",
-		"parent": "components"
-	},
-	{
-		"title": "V2",
-		"slug": "v2",
-		"markdown_source": "../packages/components/src/dropdown-menu/v2/README.md",
+		"title": "DropdownMenu",
+		"slug": "dropdown-menu",
+		"markdown_source": "../packages/components/src/dropdown-menu/README.md",
 		"parent": "components"
 	},
 	{
