+++ docs/tool/manifest.js
@@ -14,6 +14,7 @@ const componentPaths = glob( 'packages/components/src/*/**/README.md', {
 		'**/src/ui/**/README.md',
 		'packages/components/src/theme/README.md',
 		'packages/components/src/view/README.md',
+		'packages/components/src/dropdown-menu-v2/README.md',
 	],
 } );
 const packagePaths = glob( 'packages/*/package.json' )
