+++ packages/components/src/dropdown-menu/index.native.js
@@ -13,8 +13,8 @@ import { menu } from '@wordpress/icons';
 /**
  * Internal dependencies
  */
-import Button from '../../button';
-import Dropdown from '../../dropdown';
+import Button from '../button';
+import Dropdown from '../dropdown';
 
 function mergeProps( defaultProps = {}, props = {} ) {
 	const mergedProps = {
