+++ packages/components/src/dropdown-menu/index.tsx
@@ -11,9 +11,9 @@ import { menu } from '@wordpress/icons';
 /**
  * Internal dependencies
  */
-import Button from '../../button';
-import Dropdown from '../../dropdown';
-import { NavigableMenu } from '../../navigable-container';
+import Button from '../button';
+import Dropdown from '../dropdown';
+import { NavigableMenu } from '../navigable-container';
 import type { DropdownMenuProps, DropdownOption } from './types';
 
 function mergeProps<
