+++ packages/components/src/dropdown-menu/types.ts
@@ -5,11 +5,11 @@ import type { ReactNode } from 'react';
 /**
  * Internal dependencies
  */
-import type { ButtonAsButtonProps } from '../../button/types';
-import type { WordPressComponentProps } from '../../ui/context';
-import type { DropdownProps } from '../../dropdown/types';
-import type { Props as IconProps } from '../../icon';
-import type { NavigableMenuProps } from '../../navigable-container/types';
+import type { ButtonAsButtonProps } from '../button/types';
+import type { WordPressComponentProps } from '../ui/context';
+import type { DropdownProps } from '../dropdown/types';
+import type { Props as IconProps } from '../icon';
+import type { NavigableMenuProps } from '../navigable-container/types';
 
 export type DropdownOption = {
 	/**
